+++ canary/src/main/java/io/littlehorse/canary/Main.java
@@ -1,11 +1,11 @@
-package io.littlehorse;
+package io.littlehorse.canary;
 
-import io.littlehorse.common.app.BoostrapInitializationException;
-import io.littlehorse.common.app.Bootstrap;
-import io.littlehorse.common.config.CanaryConfig;
-import io.littlehorse.common.config.ConfigLoader;
-import io.littlehorse.kafka.KafkaBootstrap;
-import io.littlehorse.metronome.WorkerBootstrap;
+import io.littlehorse.canary.app.BoostrapInitializationException;
+import io.littlehorse.canary.app.Bootstrap;
+import io.littlehorse.canary.config.CanaryConfig;
+import io.littlehorse.canary.config.ConfigLoader;
+import io.littlehorse.canary.kafka.KafkaBootstrap;
+import io.littlehorse.canary.metronome.WorkerBootstrap;
 import java.io.IOException;
 import java.nio.file.Paths;
 import java.util.List;
@@ -13,7 +13,7 @@ import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
 public class Main {
-    public static void main(String[] args) throws IOException, BoostrapInitializationException {
+    public static void main(String[] args) throws IOException, BoostrapInitializationException, InterruptedException {
         CanaryConfig config = args.length > 0 ? ConfigLoader.load(Paths.get(args[0])) : ConfigLoader.load();
 
         log.info("Canary configurations: {}", config);
