+++ canary/src/main/java/io/littlehorse/canary/config/LittleHorseConfig.java
@@ -1,6 +1,6 @@
-package io.littlehorse.common.config;
+package io.littlehorse.canary.config;
 
-import static io.littlehorse.common.config.CanaryConfig.LH_CANARY_PREFIX;
+import static io.littlehorse.canary.config.CanaryConfig.LH_CANARY_PREFIX;
 import static java.util.Map.entry;
 
 import io.littlehorse.sdk.common.config.LHConfig;
