+++ canary/src/main/resources/log4j2.properties
@@ -44,7 +44,7 @@ logger.kafka.additivity = false
 
 # canary logger
 logger.canary = TRACE, canary
-logger.canary.name = io.littlehorse
+logger.canary.name = io.littlehorse.canary
 logger.canary.additivity = false
 
 # lh logger
