+++ canary/src/test/java/io/littlehorse/common/config/ConfigLoaderTest.java
@@ -3,6 +3,8 @@ package io.littlehorse.common.config;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.entry;
 
+import io.littlehorse.canary.config.CanaryConfig;
+import io.littlehorse.canary.config.ConfigLoader;
 import java.io.FileWriter;
 import java.io.IOException;
 import java.nio.file.Files;
