+++ canary/src/test/java/io/littlehorse/common/config/KafkaAdminConfigTest.java
@@ -3,6 +3,7 @@ package io.littlehorse.common.config;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.entry;
 
+import io.littlehorse.canary.config.KafkaAdminConfig;
 import java.util.Map;
 import org.junit.jupiter.api.Test;
